Switch back to pygments syntax highlighter
This seems to be what GitHub is complaining about (albeit without actually
saying that). I switched to rouge (which is apparently supported) because
pygments was failing locally, but now that I've taken out all default
content with code blocks it won't be a problem. I'll deal with it later
when I come to start putting code on the blog.